Initial commit scara version - probably not ready for public consumption
This is essentially a merge of the scara code by qharley and a more recent branch of marlin (i needed minitronics support).  I have attempted to better integrate the code of qharley so that it doesn't replace the delta code but rather lives alongside it.  I have also greatly modified many chunks of code including the core inverse kinematics calculation (theoretically it should be faster).  There is still quite some work to do so this should still be regarded as experimental.